requirements.txt
urwid
python-dotenv